readonly on displayed reviews and scrub user info
